added rsContinuedFractionGenerator

+++ Tests/TestsRosicAndRapt/Source/rapt_tests/UnitTests/Math/MiscMathUnitTests.cpp
@@ -917,7 +917,7 @@ bool testFraction()  // maybe move up
 
 
   // try to produce the continued fraction expansion of pi:
-  rsContinuedFractionGenerator<int, double> cfg(PI);
+  rsContinuedFractionGenerator<int, long double> cfg(PI);
   int N = 20;
   std::vector<int> cfe(N);
   for(int n = 0; n < N; n++)
@@ -934,6 +934,7 @@ bool testFraction()  // maybe move up
   // https://oeis.org/A002485
   // https://oeis.org/A002486
 
+  // test compiling for x86-32 - it has the 80 bit long double type
 
 
   return res;
